python_utils/requirements.txt
numpy
pyserial
matplotlib
opencv-python
mediapipe
